fix: no need migration rollback if environment is testing

+++ database/migrations/2024_07_21_200129_add_is_censored_into_posts.php
@@ -25,6 +25,9 @@ class AddIsCensoredIntoPosts extends Migration
      */
     public function down()
     {
+        if(app()->environment() === 'testing') {
+            return;
+        }
         Schema::table('posts', function (Blueprint $table) {
             $table->dropColumn('is_censored');
         });
